Add mobile checks and restructure app if view is too small
