+++ package.json
@@ -20,7 +20,6 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-redux": "^9.1.0",
-    "react-responsive": "^9.0.2",
     "reactflow": "^11.9.4",
     "redux": "^5.0.1"
   },
