+++ src/library/header/Header.tsx
@@ -15,8 +15,22 @@ const Header = ({ skyColour, buildingColour }: HeaderProps) => {
       <Flex position={"relative"} paddingTop={2} paddingLeft={2}>
         <Box h={16} w={2} bg={"scotland.brown"} />
         <Box h={10} w={20} bg={"scotland.blue"} />
-        <Box h={10} w={20} bg={"white"} clipPath={"polygon(0 5%, 0 0, 5% 0, 100% 95%, 100% 100%, 95% 100%)"} position={"absolute"} marginLeft={2}/>
-        <Box h={10} w={20} bg={"white"} clipPath={"polygon(95% 0, 100% 0, 100% 5%, 5% 100%, 0 100%, 0 95%)"} position={"absolute"} marginLeft={2}/>
+        <Box
+          h={10}
+          w={20}
+          bg={"white"}
+          clipPath={"polygon(0 5%, 0 0, 5% 0, 100% 95%, 100% 100%, 95% 100%)"}
+          position={"absolute"}
+          marginLeft={2}
+        />
+        <Box
+          h={10}
+          w={20}
+          bg={"white"}
+          clipPath={"polygon(95% 0, 100% 0, 100% 5%, 5% 100%, 0 100%, 0 95%)"}
+          position={"absolute"}
+          marginLeft={2}
+        />
       </Flex>
       <Box h={16} bg={buildingColour} clipPath={ROOF_CLIP_PATH} />
     </Box>
